Add hook scripts for dhcpcd and gnome NetworkManager
Some of the changes we make through networkmanager trigger additional
actions from other networking components, most particular DHCP. The
hook scripts provided here help keep the network settings healthy.

Note that no hook script is provided for dhclient. This is because
dhclient returns an error state when it falls back to a fixed address.
For this reason dhclient was found to be unsuited for the Bubba OS.