let GUI dynamically manage Inotify watches:
MiniDLNA apparently requires a huge number of Inotify watches, even though the numbers
it returns in the warning message don't really confirm this. While the check performed
by MiniDLNA may be wrong, we still want to get rid of the warning.

The logic added here lets the services controller increase Inotify max_user_watches
when MiniDLNA is enabled en return to default when the service is disabled.