Redirect when calling `lanupdate` and `wanupdate` with empty postdata
When you change network settings this sometimes results in a timeout
loading the web page. Reloading the page after this happens would cause
the GUI to reset the NIC to dynamic.

Also correct a tiny nuisance with user control panel showing the session
stored value for `AllowRemote` rather than the actual saved value.